Push command responder pieces to library modules
This is done mostly in preparation to command responder
with notification originator merger.
